added tileview logic and amount owed logic

+++ app/src/main/java/br/edu/scl/ifsp/ads/splitthebill/model/Participant.kt
@@ -7,7 +7,7 @@ import kotlinx.parcelize.Parcelize
 data class Participant(
     var id: Int,
     var name: String,
-    var amountSpent: Float,
+    var amountSpent: Double,
     var itemsBought: String,
 ) : Parcelable {
